Fix problems with merge
