+++ app/helpers/tutor_helper.rb
@@ -1,0 +1,2 @@
+module TutorHelper
+end
