+++ test/controllers/tutor_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class TutorControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
